fix: :bug: 20250930 イベントコード取得処理の共通化
