+++ src/app/user/signin/page.tsx
@@ -1,17 +1,17 @@
 "use client";
 
 import { useState, useEffect } from "react";
-import { checkIsVisibleUser, signIn } from "@/lib/auth";
 import { useRouter } from "next/navigation";
 import { User } from "@supabase/supabase-js";
 import { Box, Typography, Alert, Card, CardContent, Link, Stack, CircularProgress } from "@mui/material";
 import { Email, Lock } from "@mui/icons-material";
 import { CustomTextField } from "@/components/base/CustomTextField";
 import CustomButton from "@/components/base/CustomButton";
-import { GetUsersByUuidResponse } from "@/features/users/[uuid]/types";
 import { Messages } from "@/constants/messages";
 import { ApplicationErrorFactory } from "@/error/applicationError";
 import { ApplicationErrorHandler } from "@/error/errorHandler";
+import { UserUtils } from "@/utils/userUtils";
+import { signIn } from "@/lib/auth";
 
 /**
  * サインインページコンポーネント
@@ -72,37 +72,10 @@ const SignInPage: React.FC = (): React.JSX.Element => {
             setError(null);
 
             const uuid = sbUser.id;
-            const response = await fetch(`/api/users/${uuid}`);
-            if (!response.ok) {
-                throw ApplicationErrorFactory.createFromResponse(response);
-            }
-
-            const data: GetUsersByUuidResponse = (await response.json()).data as GetUsersByUuidResponse;
-
-            const user = data.user || {};
-            const attendances = user.attendances || [];
-
-            // 参加しているイベントのうち閲覧権限のあるものを開催日降順またはid降順でソート
-            attendances.filter((attendance) => checkIsVisibleUser(user, attendance)).sort((a, b) => {
-                if (a.event.startDate && b.event.startDate) {
-                    return new Date(b.event.startDate).getTime() - new Date(a.event.startDate).getTime();
-                } else {
-                    if (!a.event.startDate && !b.event.startDate) {
-                        return b.event.id - a.event.id;
-                    } else if (!a.event.startDate) {
-                        return 1;
-                    } else if (!b.event.startDate) {
-                        return -1;
-                    }
-                    return 0;
-                }
-            });
-
-            return attendances.shift()?.eventCode || "";
+            return await UserUtils.fetchEventCode(uuid);
         } catch (error) {
             const appError = ApplicationErrorFactory.normalize(error);
-            ApplicationErrorHandler.logError(appError);
-
+            ApplicationErrorHandler.logError(appError, "WARN");
             return "";
         }
     };
